requirements.txt
blinker==1.8.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gitdb==4.0.11
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pyarrow==15.0.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.3
